YT-23616: Rename raw_client to http_client commit_hash:df330f3a0c0ca36d9bcf801fd96b964f1be6383a
