+++ yt/cpp/mapreduce/client/client_reader.cpp
@@ -13,15 +13,13 @@
 #include <yt/cpp/mapreduce/http/retry_request.h>
 
 #include <yt/cpp/mapreduce/interface/config.h>
+#include <yt/cpp/mapreduce/interface/raw_client.h>
 #include <yt/cpp/mapreduce/interface/tvm.h>
 
 #include <yt/cpp/mapreduce/interface/logging/yt_log.h>
 
 #include <yt/cpp/mapreduce/io/helpers.h>
 #include <yt/cpp/mapreduce/io/yamr_table_reader.h>
-
-#include <yt/cpp/mapreduce/raw_client/raw_client.h>
-#include <yt/cpp/mapreduce/raw_client/raw_requests.h>
 
 #include <library/cpp/yson/node/serialize.h>
