+++ yt/cpp/mapreduce/client/lock.cpp
@@ -2,18 +2,16 @@
 
 #include "yt_poller.h"
 
-#include <yt/cpp/mapreduce/http/retry_request.h>
-
 #include <yt/cpp/mapreduce/common/retry_lib.h>
 
-#include <yt/cpp/mapreduce/raw_client/raw_batch_request.h>
+#include <yt/cpp/mapreduce/http/retry_request.h>
+
+#include <yt/cpp/mapreduce/interface/raw_batch_request.h>
 
 #include <util/string/builder.h>
 
 namespace NYT {
 namespace NDetail {
-
-using namespace NRawClient;
 
 ////////////////////////////////////////////////////////////////////////////////
