+++ yt/cpp/mapreduce/client/structured_table_formats.cpp
@@ -14,7 +14,7 @@
 
 #include <yt/cpp/mapreduce/library/table_schema/protobuf.h>
 
-#include <yt/cpp/mapreduce/raw_client/raw_requests.h>
+#include <yt/cpp/mapreduce/http_client/raw_requests.h>
 
 #include <library/cpp/type_info/type_info.h>
 #include <library/cpp/yson/writer.h>
