+++ yt/cpp/mapreduce/client/transaction_pinger.cpp
@@ -13,8 +13,6 @@
 #include <yt/cpp/mapreduce/http/requests.h>
 #include <yt/cpp/mapreduce/http/retry_request.h>
 
-#include <yt/cpp/mapreduce/raw_client/raw_requests.h>
-
 #include <yt/yt/core/concurrency/periodic_executor.h>
 #include <yt/yt/core/concurrency/poller.h>
 #include <yt/yt/core/concurrency/scheduler_api.h>
