+++ yt/cpp/mapreduce/client/ya.make
@@ -39,9 +39,9 @@ PEERDIR(
     library/cpp/yson
     yt/cpp/mapreduce/common
     yt/cpp/mapreduce/http
+    yt/cpp/mapreduce/http_client
     yt/cpp/mapreduce/interface
     yt/cpp/mapreduce/io
-    yt/cpp/mapreduce/raw_client
 )
 
 PEERDIR(
@@ -61,7 +61,7 @@ IF (BUILD_TYPE == "PROFILE")
     ELSE()
         SRCS(
             job_profiler.cpp
-        )    
+        )
     ENDIF()
 ELSE()
     SRCS(
