+++ yt/cpp/mapreduce/http_client/raw_batch_request.cpp
