+++ yt/cpp/mapreduce/http_client/ut/raw_batch_request_ut.cpp
@@ -1,4 +1,4 @@
-#include <yt/cpp/mapreduce/raw_client/raw_batch_request.h>
+#include <yt/cpp/mapreduce/http_client/raw_batch_request.h>
 
 #include <yt/cpp/mapreduce/http/context.h>
 #include <yt/cpp/mapreduce/interface/client_method_options.h>
