+++ yt/cpp/mapreduce/http_client/ut/raw_requests_ut.cpp
@@ -1,4 +1,4 @@
-#include <yt/cpp/mapreduce/raw_client/raw_requests.h>
+#include <yt/cpp/mapreduce/http_client/raw_requests.h>
 
 #include <library/cpp/yson/node/node_io.h>
