+++ yt/cpp/mapreduce/http_client/ya.make
