+++ yt/cpp/mapreduce/io/yamr_table_reader.cpp
@@ -2,7 +2,6 @@
 
 #include <yt/cpp/mapreduce/common/helpers.h>
 #include <yt/cpp/mapreduce/common/retry_lib.h>
-#include <yt/cpp/mapreduce/raw_client/raw_requests.h>
 
 ////////////////////////////////////////////////////////////////////////////////
 
@@ -15,8 +14,6 @@ static void CheckedSkip(IInputStream* input, size_t byteCount)
 ////////////////////////////////////////////////////////////////////////////////
 
 namespace NYT {
-
-using namespace NYT::NDetail::NRawClient;
 
 ////////////////////////////////////////////////////////////////////////////////
